requirements.txt
pyqt4
pymongo
tesseract
re
numpy
os
opencv-python
pdf2image
pillow